Issue 9 - Extracting area form data interface
