+++ app/formData.ts
@@ -1,0 +1,4 @@
+import { z } from "zod";
+import { areaSchema } from "./validationSchema";
+
+export type AreaFormData = z.infer<typeof areaSchema>;
